feat: 처음에 데이터 심는 과정 추가
